Analyse Ablauf aufgebaut

+++ src/main/java/de/mankianer/gruppe5/model/analyse/Analyse.java
@@ -12,7 +12,7 @@ public class Analyse implements Serializable{
 	 */
 	private static final long serialVersionUID = 1L;
 	
-	private long tweetID;
+	private int tweetID;
 
 	public Analyse(String result) {
 		this.result = result;
